Moved all dependencies into libs package
