+++ pom.xml
@@ -100,11 +100,11 @@
 <!--			<artifactId>samplesvc-api</artifactId>-->
 <!--			<version>10.49.19</version>-->
 <!--		</dependency>-->
-		<dependency>
-			<groupId>uk.gov.ons.ctp.product</groupId>
-			<artifactId>reportmodule</artifactId>
-			<version>10.49.0</version>
-		</dependency>
+<!--		<dependency>-->
+<!--			<groupId>uk.gov.ons.ctp.product</groupId>-->
+<!--			<artifactId>reportmodule</artifactId>-->
+<!--			<version>10.49.0</version>-->
+<!--		</dependency>-->
 		<dependency>
 			<groupId>uk.gov.ons.ctp.common</groupId>
 			<artifactId>framework</artifactId>
