+++ src/main/java/uk/gov/ons/ctp/response/libs/party/representation/PartyDTO.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.response.sample.party;
+package uk.gov.ons.ctp.response.libs.party.representation;
 
 import java.util.List;
 import lombok.AccessLevel;
