+++ src/main/java/uk/gov/ons/ctp/response/libs/sample/validation/SampleUnitBase.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.response.sample.validation;
+package uk.gov.ons.ctp.response.libs.sample.validation;
 
 import java.util.UUID;
 import lombok.AllArgsConstructor;
