+++ src/main/java/uk/gov/ons/ctp/response/sample/mapper/SampleUnitMapper.java
@@ -2,7 +2,7 @@ package uk.gov.ons.ctp.response.sample.mapper;
 
 import ma.glasnost.orika.CustomMapper;
 import ma.glasnost.orika.MappingContext;
-import uk.gov.ons.ctp.response.libs.SampleUnit;
+import uk.gov.ons.ctp.response.libs.sample.definition.SampleUnit;
 
 public class SampleUnitMapper
     extends CustomMapper<uk.gov.ons.ctp.response.sample.domain.model.SampleUnit, SampleUnit> {
