+++ src/main/java/uk/gov/ons/ctp/response/sample/service/PartySvcClientService.java
@@ -16,10 +16,10 @@ import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponents;
 import uk.gov.ons.ctp.common.rest.RestUtility;
-import uk.gov.ons.ctp.response.libs.PartyCreationRequestDTO;
+import uk.gov.ons.ctp.response.libs.party.definition.PartyCreationRequestDTO;
+import uk.gov.ons.ctp.response.libs.party.representation.PartyDTO;
 import uk.gov.ons.ctp.response.sample.config.AppConfig;
 import uk.gov.ons.ctp.response.sample.message.EventPublisher;
-import uk.gov.ons.ctp.response.sample.party.PartyDTO;
 
 @Service
 public class PartySvcClientService {
