+++ src/main/resources/springintegration/party-creation-in.xml
@@ -36,7 +36,7 @@
    <bean id="partyInMarshaller" class="org.springframework.oxm.jaxb.Jaxb2Marshaller">
       <property name="contextPath" value="uk.gov.ons.ctp.response.party.definition" />
    </bean>
-   <int:channel id="partyTransformed" datatype="uk.gov.ons.ctp.response.libs.PartyCreationRequestDTO" />
+   <int:channel id="partyTransformed" datatype="uk.gov.ons.ctp.response.libs.party.definition.PartyCreationRequestDTO" />
    <int-xml:unmarshalling-transformer
            input-channel="partyXmlValid" output-channel="partyTransformed"
            unmarshaller="partyInMarshaller" />
