+++ src/test/java/uk/gov/ons/ctp/response/sample/validation/BusinessSurveySample.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import net.sourceforge.cobertura.CoverageIgnore;
+import uk.gov.ons.ctp.response.libs.sample.validation.BusinessSampleUnit;
 
 @CoverageIgnore
 @Data
